task-tracker/requirements.txt
fastapi[all]
tortoise-orm[aiosqlite]